Wip blog backend

+++ blog/src/main/java/com/blog/controller/PostController.java
@@ -2,6 +2,7 @@ package com.blog.controller;
 
 import com.blog.database.Database;
 import com.blog.exception.BlogException;
+import com.blog.exception.InvalidPermissionException;
 import com.blog.model.Post;
 import com.blog.model.User;
 import com.blog.model.UserLevel;
@@ -43,6 +44,8 @@ public class PostController {
         try {
             createPost(new JSONObject(input));
             return ResponseEntity.ok(getPost(new JSONObject(input)).toString());
+        } catch (InvalidPermissionException e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.FORBIDDEN);
         } catch (Exception e) {
             return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
         }
@@ -54,6 +57,8 @@ public class PostController {
         try {
             deletePost(new JSONObject(input));
             return ResponseEntity.ok().build();
+        } catch (InvalidPermissionException e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.FORBIDDEN);
         } catch (Exception e) {
             return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
         }
@@ -65,6 +70,8 @@ public class PostController {
         try {
             editPost(new JSONObject(input));
             return ResponseEntity.ok(getPost(new JSONObject(input)).toString());
+        } catch (InvalidPermissionException e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.FORBIDDEN);
         } catch (Exception e) {
             return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
         }
@@ -199,10 +206,10 @@ public class PostController {
      *
      * @param input A JSON containing the following key-value pairs:
      *              {
-     *              "authorID": int,          // The author of the post.
-     *              "title":    String        // The title of the post.
-     *              "content":  String        // The content of the post.
-     *              "allowcomments": boolean  // Whether to allow comments
+     *              "authorID":      int,     // The author of the post.
+     *              "title":         String   // The title of the post.
+     *              "content":       String   // The content of the post.
+     *              "allowComments": boolean  // Whether to allow comments
      *              }
      * @throws BlogException
      */
@@ -224,8 +231,13 @@ public class PostController {
             throw new BlogException("JSON object received is null. \n" + e.getMessage());
         }
 
-        // Check whether author has permission to make a post
-        validatePermission(authorID);
+        // Retrieve the user
+        User user = new User(authorID);
+
+        // Check whether the user has permission to make a post. // TODO: change to MEMBER
+        if (UserLevel.READER.compareTo(user.getUserLevel()) < 0) {
+            throw new InvalidPermissionException("User does not have the necessary permission to make a post.");
+        }
 
         // Validate the data
         Post.validateTitle(title);
@@ -256,13 +268,22 @@ public class PostController {
      *
      * @param input A JSON containing the following key-value pairs:
      *              {
-     *              "postID":    int,  // The post to delete.
+     *              "postID": int,  // The post to delete.
+     *              "userID": int,  // The user attempting to delete.
      *              }
      * @throws BlogException
      */
     private static void deletePost(JSONObject input) throws BlogException {
         // Retrieve the post
         Post post = retrievePost(input);
+
+        // Retrieve the user
+        User user = UserController.retrieveUser(input);
+
+        // Check whether user has permission to delete post
+        if (post.getAuthorID() != user.getUserID() && UserLevel.ADMIN.compareTo(user.getUserLevel()) < 0) {
+            throw new InvalidPermissionException("User does not have the necessary permission to delete this post.");
+        }
 
         // Delete post in database
         Database.delete(post);
@@ -276,7 +297,8 @@ public class PostController {
      *              "postID":        int,     // The post to edit.
      *              "title":         String   // The new title of the post.
      *              "content":       String   // The new content of the post.
-     *              "allowcomments": boolean  // Whether to allow comments
+     *              "allowComments": boolean  // Whether to allow comments
+     *              "userID":        int,     // The user attempting to edit.
      *              }
      * @throws BlogException
      */
@@ -302,6 +324,14 @@ public class PostController {
 
         // Retrieve the post
         Post post = retrievePost(input);
+
+        // Retrieve the user
+        User user = UserController.retrieveUser(input);
+
+        // Check whether user has permission to delete post
+        if (post.getAuthorID() != user.getUserID() && UserLevel.ADMIN.compareTo(user.getUserLevel()) < 0) {
+            throw new InvalidPermissionException("User does not have the necessary permission to edit this post.");
+        }
 
         // Apply edit to post
         post.setTitle(title);
@@ -397,21 +427,5 @@ public class PostController {
 
         // Return the retrieved post
         return new Post(postID);
-    }
-
-    /**
-     * Validates whether the user has the necessary permissions to make a post.
-     *
-     * @param userID The user to validate.
-     * @throws BlogException
-     */
-    private static void validatePermission(int userID) throws BlogException {
-        // Retrieve the user
-        User user = new User(userID);
-
-        // Check whether the user has UserLevel of at least UserLevel.CONTRIBUTOR
-        if (UserLevel.CONTRIBUTOR.compareTo(user.getUserLevel()) < 0) {
-            throw new BlogException("User does not have the necessary permission to make a POST.");
-        }
     }
 }
